Possible fix for #345
I say "possible" because we don't have a way to reproduce it. Following
everything I've been able to find tho, seems like fragments are part of
the culprit (real culprit could be a hardware problem, and this is only
reproducible in chrome browsers... so pretty complicated to reproduce it
on our own). There's some plugin, or even the same browser, doing some
replacement at some point directly to a <Fragment /> element, making
the app crash, because fragments aren't really DOM elements.

For that reason, the only two fragments that could be affecting this
behavior have been changed to `<div />`.

closes #345

at least until we cannot reproduce it again 🤷